WahWahBase in lib-builtin-effects
